Add magician MP UI and magician can now use MP to operate and put the origin HP UI to the button left.

+++ Assets/Scripts/Magician/MagicianCallAllPlayerToShakeDevice.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class MagicianCallAllPlayerToShakeDevice : Photon.MonoBehaviour {
 
-	public float CostMP = 50;
+	public int CostMP = 50;
 
 	public void CastSpell()
 	{
